requirements.txt
numpy==1.19.0
pandas==1.0.2
Pillow==7.1.2
plotly==4.9.0
pydeck==0.4.0b2
streamlit==0.65.2
